linux-gen: modular-framework: move pool subsystem header to a public folder
Move pool subsystem header to a public folder.
